# api/docs/swagger.yaml
basePath: /local-eats
definitions:
...
  payment.NewPayment:
    properties:
      card_number:
        type: string
      cvv:
        type: string
      expiry_date:
        type: string
      order_id:
        type: string
      payment_method:
        type: string
    type: object
  payment.PaymentDetails:
    properties:
      amount:
        type: number
      card_number:
        type: string
      created_at:
        type: string
      cvv:
        type: string
      expiry_date:
        type: string
      id:
        type: string
      method:
        type: string
      order_id:
        type: string
      status:
        type: string
      transaction_id:
        type: string
    type: object
  review.NewReview:
    properties:
      comment:
        type: string
      order_id:
        type: string
      rating:
        type: number
    type: object
  review.NewReviewResp:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      kitchen_id:
        type: string
      order_id:
        type: string
      rating:
        type: number
      user_id:
        type: string
    type: object
  review.ReviewDetails:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: number
      user_name:
        type: string
    type: object
  review.Reviews:
    properties:
      average_rating:
        type: number
      limit:
        type: integer
      page:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/review.ReviewDetails'
        type: array
      total:
        type: integer
    type: object
  user.Details:
    properties:
      address:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
  user.NewInfoNoID:
    properties:
      address:
        type: string
      full_name:
        type: string
      phone_number:
        type: string
    type: object
  user.Profile:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
  user.Void:
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway of Local Eats
  title: Local Eats
  version: "1.0"
paths:
...
  /kitchens/{id}/reviews:
    get:
      description: Gets reviews from database
      parameters:
      - description: Kitchen ID
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.Reviews'
        "400":
          description: Invalid review data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets reviews
      tags:
      - review
...
  /payments:
    post:
      description: Inserts a new payment into database
      parameters:
      - description: Payment info
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/payment.NewPayment'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.NewPayment'
        "400":
          description: Invalid payment data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a payment
      tags:
      - payment
  /payments/{id}:
    get:
      description: Retrieves payment info from database
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.PaymentDetails'
        "400":
          description: Invalid payment ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets a payment
      tags:
      - payment
  /reviews:
    post:
      description: Inserts a new review into database
      parameters:
      - description: Review info
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/review.NewReview'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.NewReviewResp'
        "400":
          description: Invalid review data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a review
      tags:
      - review
  /users/{id}:
    delete:
      description: Deletes user from database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Void'
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes a user
      tags:
      - user
    get:
      description: Retrieves user info from database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Profile'
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets a user
      tags:
      - user
    put:
      description: Updates user info in database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.NewInfoNoID'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Details'
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates a user
      tags:
      - user
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"